requirements.txt
line-bot-sdk==3.10
flask==3.0.0